# packages/sui-media-api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/uploads/initiate": {
      "post": {
        "operationId": "UploadsController_initiateUpload",
        "summary": "Initiate multipart upload",
        "description": "Start a new resumable multipart upload. Returns presigned URLs for the first batch of parts.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload initiated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiateUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Uploads"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/uploads/{sessionId}/status": {
      "get": {
        "operationId": "UploadsController_getUploadStatus",
        "summary": "Get upload status",
        "description": "Get the current status of an upload session, including pending parts and fresh presigned URLs for resume.",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Upload session ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "includeUrls",
            "required": false,
            "in": "query",
            "description": "Include fresh presigned URLs for pending parts (default: true)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Upload status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadStatusResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to access this session"
          },
          "404": {
            "description": "Upload session not found"
          }
        },
        "tags": [
          "Uploads"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/uploads/{sessionId}/sync": {
      "post": {
        "operationId": "UploadsController_syncUpload",
        "summary": "Sync upload with S3",
        "description": "Synchronize the upload session with actual S3 state. Use this when resuming an upload to ensure consistency.",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Upload session ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Upload synced and status retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadStatusResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Upload session not found"
          },
          "409": {
            "description": "Upload session expired or already completed"
          }
        },
        "tags": [
          "Uploads"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/uploads/{sessionId}/urls": {
      "post": {
        "operationId": "UploadsController_getMoreUrls",
        "summary": "Get more presigned URLs",
        "description": "Request additional presigned URLs for specific parts (max 50 per request).",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Upload session ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetMoreUrlsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presigned URLs generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PresignedUrlDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid part numbers or too many requested"
          },
          "409": {
            "description": "Upload session is not active"
          }
        },
        "tags": [
          "Uploads"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/uploads/{sessionId}/parts/{partNumber}/complete": {
      "post": {
        "operationId": "UploadsController_markPartCompleted",
        "summary": "Mark part as completed",
        "description": "Confirm that a part has been successfully uploaded to S3 with its ETag.",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Upload session ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "partNumber",
            "required": true,
            "in": "path",
            "description": "Part number (1-based)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartCompletionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Part marked as completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartCompletionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid part number or ETag"
          },
          "409": {
            "description": "Upload session is not active"
          }
        },
        "tags": [
          "Uploads"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/uploads/{sessionId}/complete": {
      "post": {
        "operationId": "UploadsController_completeUpload",
        "summary": "Complete upload",
        "description": "Finalize the multipart upload after all parts have been uploaded. Creates the media document.",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Upload session ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Upload completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Not all parts have been uploaded"
          },
          "409": {
            "description": "Upload already completed or aborted"
          }
        },
        "tags": [
          "Uploads"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/uploads/{sessionId}": {
      "delete": {
        "operationId": "UploadsController_abortUpload",
        "summary": "Abort upload",
        "description": "Cancel an in-progress upload and clean up S3 resources.",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Upload session ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Upload aborted successfully"
          },
          "404": {
            "description": "Upload session not found"
          },
          "409": {
            "description": "Cannot abort a completed upload"
          }
        },
        "tags": [
          "Uploads"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/uploads/active": {
      "get": {
        "operationId": "UploadsController_getActiveUploads",
        "summary": "List active uploads",
        "description": "Get all active (resumable) uploads for the current user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Active uploads retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActiveUploadsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Uploads"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Stoked UI Media API",
    "description": "REST API for Stoked UI Media Components - Upload, manage, and process media files with support for images, videos, and albums. Features include multipart uploads, metadata extraction, thumbnail generation, and comprehensive media management.",
    "version": "0.1.0",
    "contact": {
      "name": "Brian Stoker",
      "url": "https://github.com/stoked-ui/sui",
      "email": "brian@stoked-ui.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    },
    {
      "url": "https://api.stoked-ui.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "InitiateUploadDto": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "description": "Original filename of the file being uploaded",
            "example": "my-video.mp4"
          },
          "mimeType": {
            "type": "string",
            "description": "MIME type of the file",
            "example": "video/mp4"
          },
          "totalSize": {
            "type": "number",
            "description": "Total file size in bytes",
            "example": 104857600,
            "minimum": 1
          },
          "hash": {
            "type": "string",
            "description": "SHA-256 hash of the file content for deduplication (base64 encoded)",
            "example": "Uonfc331cyb83SJZevsfrA=="
          },
          "chunkSize": {
            "type": "number",
            "description": "Chunk size in bytes (default: 10MB, min: 5MB, max: 100MB)",
            "example": 10485760,
            "minimum": 5242880,
            "maximum": 104857600
          }
        },
        "required": [
          "filename",
          "mimeType",
          "totalSize"
        ]
      },
      "PresignedUrlDto": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "number",
            "description": "Part number (1-based)",
            "example": 1
          },
          "url": {
            "type": "string",
            "description": "Presigned PUT URL for uploading the part directly to S3",
            "example": "https://bucket.s3.amazonaws.com/key?X-Amz-..."
          }
        },
        "required": [
          "partNumber",
          "url"
        ]
      },
      "InitiateUploadResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Upload session ID (use this for all subsequent operations)",
            "example": "507f1f77bcf86cd799439011"
          },
          "uploadId": {
            "type": "string",
            "description": "S3 multipart upload ID",
            "example": "2~..."
          },
          "presignedUrls": {
            "description": "Presigned URLs for the first batch of parts",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PresignedUrlDto"
            }
          },
          "totalParts": {
            "type": "number",
            "description": "Total number of parts to upload",
            "example": 10
          },
          "chunkSize": {
            "type": "number",
            "description": "Chunk size in bytes",
            "example": 10485760
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the upload session expires",
            "example": "2024-01-15T00:00:00.000Z"
          },
          "existingMedia": {
            "type": "object",
            "description": "If set, a duplicate file was found and no upload is needed"
          }
        },
        "required": [
          "sessionId",
          "uploadId",
          "presignedUrls",
          "totalParts",
          "chunkSize",
          "expiresAt"
        ]
      },
      "UploadStatusResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Upload session ID",
            "example": "507f1f77bcf86cd799439011"
          },
          "status": {
            "type": "string",
            "description": "Current status of the upload",
            "enum": [
              "pending",
              "in_progress",
              "completed",
              "aborted",
              "expired"
            ],
            "example": "in_progress"
          },
          "filename": {
            "type": "string",
            "description": "Original filename",
            "example": "my-video.mp4"
          },
          "totalSize": {
            "type": "number",
            "description": "Total file size in bytes",
            "example": 104857600
          },
          "totalParts": {
            "type": "number",
            "description": "Total number of parts",
            "example": 10
          },
          "completedParts": {
            "type": "number",
            "description": "Number of completed parts",
            "example": 5
          },
          "progress": {
            "type": "number",
            "description": "Upload progress percentage (0-100)",
            "example": 50
          },
          "pendingPartNumbers": {
            "description": "Part numbers that still need to be uploaded",
            "example": [
              6,
              7,
              8,
              9,
              10
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "presignedUrls": {
            "description": "Fresh presigned URLs for pending parts (up to 50)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PresignedUrlDto"
            }
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the upload session expires",
            "example": "2024-01-15T00:00:00.000Z"
          }
        },
        "required": [
          "sessionId",
          "status",
          "filename",
          "totalSize",
          "totalParts",
          "completedParts",
          "progress",
          "pendingPartNumbers",
          "expiresAt"
        ]
      },
      "GetMoreUrlsDto": {
        "type": "object",
        "properties": {
          "partNumbers": {
            "description": "Array of part numbers to get presigned URLs for (1-based)",
            "example": [
              5,
              6,
              7,
              8,
              9,
              10
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "partNumbers"
        ]
      },
      "PartCompletionDto": {
        "type": "object",
        "properties": {
          "etag": {
            "type": "string",
            "description": "ETag returned by S3 after uploading the part",
            "example": "\"d41d8cd98f00b204e9800998ecf8427e\""
          }
        },
        "required": [
          "etag"
        ]
      },
      "PartCompletionResponseDto": {
        "type": "object",
        "properties": {
          "completedParts": {
            "type": "number",
            "description": "Number of completed parts",
            "example": 5
          },
          "totalParts": {
            "type": "number",
            "description": "Total number of parts",
            "example": 10
          },
          "progress": {
            "type": "number",
            "description": "Upload progress percentage (0-100)",
            "example": 50
          }
        },
        "required": [
          "completedParts",
          "totalParts",
          "progress"
        ]
      },
      "CompleteUploadResponseDto": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string",
            "description": "ID of the created media document",
            "example": "507f1f77bcf86cd799439011"
          },
          "mediaType": {
            "type": "string",
            "description": "Type of media created",
            "enum": [
              "video",
              "image"
            ],
            "example": "video"
          }
        },
        "required": [
          "mediaId",
          "mediaType"
        ]
      },
      "ActiveUploadDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "description": "Upload session ID",
            "example": "507f1f77bcf86cd799439011"
          },
          "filename": {
            "type": "string",
            "description": "Original filename",
            "example": "my-video.mp4"
          },
          "totalSize": {
            "type": "number",
            "description": "Total file size in bytes",
            "example": 104857600
          },
          "progress": {
            "type": "number",
            "description": "Upload progress percentage (0-100)",
            "example": 50
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the upload was started",
            "example": "2024-01-08T00:00:00.000Z"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the upload session expires",
            "example": "2024-01-15T00:00:00.000Z"
          }
        },
        "required": [
          "sessionId",
          "filename",
          "totalSize",
          "progress",
          "createdAt",
          "expiresAt"
        ]
      },
      "ActiveUploadsResponseDto": {
        "type": "object",
        "properties": {
          "uploads": {
            "description": "List of active uploads that can be resumed",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActiveUploadDto"
            }
          }
        },
        "required": [
          "uploads"
        ]
      }
    }
  }
}